+++ cmd/kubeless/route.go
@@ -14,34 +14,24 @@ limitations under the License.
 package main
 
 import (
-	"github.com/kubeless/kubeless/pkg/utils"
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
-var ingressDeleteCmd = &cobra.Command{
-	Use:   "delete <name>",
-	Short: "delete a route from Kubeless",
-	Long:  `delete a route from Kubeless`,
+var ingressCmd = &cobra.Command{
+	Use:   "route SUBCOMMAND",
+	Short: "manage route to function on Kubeless",
+	Long:  `route command allows user to list, create, delete routing rule for function on Kubeless`,
 	Run: func(cmd *cobra.Command, args []string) {
-		if len(args) != 1 {
-			logrus.Fatal("Need exactly one argument - route name")
-		}
-		ingName := args[0]
-
-		ns, err := cmd.Flags().GetString("namespace")
-		if err != nil {
-			logrus.Fatal(err)
-		}
-		if ns == "" {
-			ns = utils.GetDefaultNamespace()
-		}
-
-		client := utils.GetClientOutOfCluster()
-
-		err = utils.DeleteIngress(client, ingName, ns)
-		if err != nil {
-			logrus.Fatal(err)
-		}
+		cmd.Help()
 	},
+}
+
+func init() {
+	cmds := []*cobra.Command{routeCreateCmd, routeListCmd, routeDeleteCmd}
+
+	for _, cmd := range cmds {
+		ingressCmd.AddCommand(cmd)
+		cmd.Flags().StringP("namespace", "n", "", "Specify namespace for the route")
+
+	}
 }
